# swagger-spec.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ghar Ka Sathi Production APIs",
    "version": "1.0.0",
    "description": "Comprehensive interactive API documentation for customer, expert, and admin application workflows. Built to support 80+ endpoints."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080/api/v1",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/services/search": {
      "get": {
        "tags": ["5. Home Dashboard"],
        "summary": "Search Services by Keyword",
        "parameters": [
          { "name": "keyword", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Match list" }
        }
      }
    },
...
    "/categories/{id}": {
      "get": {
        "tags": ["6. Categories & Services Catalog"],
        "summary": "Get Category Details",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Details" }
        }
      }
    },
    "/categories/{id}/services": {
      "get": {
        "tags": ["6. Categories & Services Catalog"],
        "summary": "Get Services inside Category",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Services list" }
        }
      }
    },
    "/categories/service/details/{id}": {
      "get": {
        "tags": ["7. Service Details"],
        "summary": "Get Service Details & FAQs",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Detailed specs, inclusions, exclusions" }
        }
      }
    },
    "/categories/service/estimate": {
      "post": {
        "tags": ["7. Service Details"],
        "summary": "Estimate price billing bill",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "serviceId": { "type": "string", "example": "sub_pl_leak" },
                  "coupon": { "type": "string", "example": "SGSAVE30" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Dynamic calculated bill values" }
        }
      }
    },
...
    "/payment/create": {
      "post": {
        "tags": ["15. Payments Settling"],
        "summary": "Create Razorpay Payment Intent",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookingId": { "type": "string", "example": "bk_123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Payment Intent ID generated" }
        }
      }
    },
    "/payment/verify": {
      "post": {
        "tags": ["15. Payments Settling"],
        "summary": "Verify payment transaction capture",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "bookingId": { "type": "string", "example": "bk_123" },
                  "paymentIntentId": { "type": "string", "example": "pi_123" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Payment verified" }
        }
      }
    },
    "/payment/{id}": {
      "get": {
        "tags": ["15. Payments Settling"],
        "summary": "Get transaction status by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Status" }
        }
      }
    },
    "/payment/{id}/receipt": {
      "get": {
        "tags": ["15. Payments Settling"],
        "summary": "Get digital receipt detail parameters",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Receipt download" }
        }
      }
    },
...
    "/notifications/{id}": {
      "patch": {
        "tags": ["18. Notifications Alert Log"],
        "summary": "Mark notification as read",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Read status set" }
        }
      },
      "delete": {
        "tags": ["18. Notifications Alert Log"],
        "summary": "Delete Notification log",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Deleted log" }
        }
      }
    },
...
    "/sessions/{id}": {
      "delete": {
        "tags": ["24. Device sessions & notifications"],
        "summary": "Terminate device session (Remote logout)",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": { "description": "Session terminated" }
        }
      }
    }
  }
}